refactor(api): reorganize dependency injection structure
This commit includes several structural changes and improvements to the Express application
architecture:

DEPENDENCY INJECTION RESTRUCTURING:
- Move DI container from /config/inversify to /infrastructure/di-container
- Maintain the same container configuration with single instance scope
- Update all imports to reference the new DI container location

SERVICE PROVIDER CHANGES:
- Move service provider to presentation layer for better separation of concerns
- Centralize dependency access through the service provider in presentation directory
- Update all service imports throughout the codebase

USER API ENDPOINT IMPROVEMENTS:
- Finalize user creation endpoint implementation
- Remove debugging console.log statements
- Complete express adapter return response in user route

CODE ORGANIZATION:
- Update all import paths for consistency with new architecture
