--- apps/express-app/src/config/inversify/index.ts
@@ -1,3 +1,0 @@
-export * from "./container";
-export * from "./service-provider";
-export * from "./types";
