+++ apps/express-app/src/presentation/express/routes/user.ts
@@ -2,8 +2,8 @@ import type { Request, Response } from "express";
 
 import { Router } from "express";
 
-import { createUserController } from "@/config/inversify";
 import { expressAdapter } from "@/presentation/adapters/express";
+import { createUserController } from "@/presentation/service-provider";
 
 /**
  * Router for handling user-related routes.
@@ -13,11 +13,10 @@ const userRoutes = Router();
 /**
  * Endpoint to create a new user.
  */
-userRoutes.post("/", async (request: Request, _response: Response) => {
+userRoutes.post("/", async (request: Request, response: Response) => {
   const adapter = await expressAdapter(request, createUserController);
 
-  console.log("adapter", adapter);
-  // return response.status(adapter.statusCode).json(adapter.body);
+  response.status(adapter.statusCode).json(adapter.body);
 });
 
 export { userRoutes };
