+++ apps/express-app/src/presentation/http/controllers/user/create-user.ts
@@ -3,7 +3,7 @@ import { inject, injectable } from "inversify";
 import type { ICreateUserUseCase } from "@/application/use-cases/user";
 import type { ICreateUserReqDTO, IUserOutReqDTO } from "@/domain/user/dtos";
 
-import { TYPES } from "@/config/inversify";
+import { TYPES } from "@/infrastructure/di-container/types";
 
 import type { IHttpRequest, IHttpResponse } from "../../helpers/interfaces";
 import type { IHttpController } from "../controller.interface";
