+++ apps/express-app/src/presentation/service-provider/index.ts
@@ -4,8 +4,8 @@ import "reflect-metadata";
 import type { ILogger } from "@/application/providers";
 import type { ICreateUserController } from "@/presentation/http/controllers/user";
 
-import { container } from "./container";
-import { TYPES } from "./types";
+import { container } from "@/infrastructure/di-container/container";
+import { TYPES } from "@/infrastructure/di-container/types";
 
 export const logger = container.get<ILogger>(TYPES.Logger);
